Refactoring & fixing

+++ user-service/src/main/java/com/example/userservice/repository/AccountRepository.java
@@ -2,14 +2,21 @@ package com.example.userservice.repository;
 
 import com.example.userservice.model.Account;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountRepository {
     void save(Account check);
+
     void delete(Account check);
-    void updateUserAndCheck(Long user_id,Long check_id);
+
+    void updateUserAndCheck(Long user_id, Long check_id);
+
     List<Account> getCheksByUserEmail(String email);
+
     Account getCheckByIdAndUserEmail(Long check_id, String email);
-    boolean updateEnableByIdAndUserEmail(Long check_id,String email,boolean bol);
-    void changeSumById(Long chek_id, Double sum);
+
+    boolean updateEnableByIdAndUserEmail(Long check_id, String email, boolean bol);
+
+    void changeSumById(Long chek_id, BigDecimal sum);
 }
